+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -236,5 +236,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-22T12:01:38.246Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-22T14:20:40.492Z")
     r"""The date and time the resource was last updated"""