+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5RMTBWDDPECVW6K13JDBH8N"
+        "01K5RWRYR8S2EYS00ABE811GS4"
     )
     r"""The id of the resource"""