+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -522,7 +522,7 @@ class Twelve(BaseModel):
     http: HTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5RMTBRS4MBMNJD655K2MPSX"
+        "01K5RWRYMC1SRKM5MT11ATVJHN"
     )
 
     status: Optional[ToolsStatus] = "live"