+++ packages/orq-rc/RELEASES.md
@@ -2839,3 +2839,13 @@ Based on:
 - [python v3.13.0-rc.25] packages/orq-rc
 ### Releases
 - [PyPI v3.13.0-rc.25] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.25 - packages/orq-rc
+
+## 2025-09-22 14:22:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.622.1 (2.709.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.13.0-rc.26] packages/orq-rc
+### Releases
+- [PyPI v3.13.0-rc.26] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.26 - packages/orq-rc